BUGFIX (empty descriptor) + Removed the unnecessary basic_rules
Dealt with a bug occurring when a set of rules containing an empty array or object.
Also cleaned the gulpfile and the settings.json to remove all traces of the previous replacement rules (the 'basic' one)
NOTE: it's still possible to describre rather 'basic' replacements, but now you don't need a separate 'basic_rules' property to declare them.